requirements.txt
PythonTurtle
requests
pandas
pyarrow
python-dotenv
pathlib